added data on orgchart personnel
