+++ templates/home/unauthed/orgchart.html
@@ -69,7 +69,7 @@
                         <div class="avatar-container">
                             <img src="{{ person.image.url }}" alt="{{ person.name }}" class="avatar">
                         </div>
-                        <h5 class="card-title">{{ person.name }}</h5>
+                        <h5 class="card-title text-center">{{ person.name }}</h5>
                         <p class="text-muted mb-0">{{ person.title }}</p>
                     </div>
                 </div>
