+++ templates/templatetags/readMe.txt
@@ -1,0 +1,4 @@
+Custom filter to be able to access _meta attributes of model instances.
+Specifically used in home/home_authed.html
+
+
